lambda-nano-banana/requirements.txt
google-genai>=0.3.0
pillow>=10.0.0
boto3>=1.26.0
python-dotenv>=1.0.0
